aiml/pyproject.toml
[project]
name = "aiml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.1",
    "flask-jwt-extended>=4.7.1",
    "flask>=3.1.2",
    "google-adk>=1.21.0",
    "google-genai>=1.56.0",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
]
